Change logic to always redraw path avoiding scaling it keeping a single source of truth for the shape draw
